AnimalView, a builder, emptyCat, and more
Created an AnimalInfoView, with a builder to create them that takes a
Cat as an argument. Needs to be tested. Created a tested static method
emptyCat() that returns an empty Cat as a placeholder for a cat. Made a
link to the MainController for the add new cat button to make empty cats
in the animal view. Added getFrame on the MainView. Holding on
connecting animalView to the GUI until that area gets more work.