Moved things, hooked up AnimalInfoViewBuilder
Moved the database classes to their own packages, tweaked a couple null
returns to emptyCat() (this will ensure we don't get
NullPointerExceptions when we want to return a "null" Cat), and hooked
up the AnimalInfoViewBuilder to the view. Next things are to clean
up/flesh out the AnimalInfoView and its layout and to hook up the
database.