Added medical history.
Added the necessary medical history information, including the
additional fields found on the medical form. The medical history is a
list of string for simplicity's sake, and the other new fields are just
booleans until we hear back from NBAR on what they usually fill in for
Rabies, FiV/FeLeuk, etc.