Started setting up resources
Added the "resource provider" object. This allows us to access any files
that are added to the project via the "resources" package. These files
will be packaged alongside a JAR deployment. I also added the PDFBox
library to the 3rd party folder and classpaths so we can start
investigating PDF auto-fill features.