Cleaned up search box interaction
Set up the search bar text field so that it will clear out the "Search
for a cat..." text when clicked on or highlight the existing filter. For
some reason the box will still get default focus even though I tried to
force it elsewhere, so until we figure that out it will still need to be
cleared manually on occasion. I also added the ability to press Enter
while editing the field to perform the search instead of having to click
the button.