Started work on cat history
Started work on the cat history form: both its model object and the
view-controller.  There is a ton of information, so I made a subpackage
for the cat history model object and was toying with ways to make it
package-protected, but I don't think it's going to work that way. Note
that I didn't change the functionality of MainController: it was merely
a victim of my autoformatting preferences.