AnimalInfoView refactoring
Merged functionality of edit and save buttons. Changed Save button to
View History button (Might change to view cat history). Started work on
the LoginHandler. It will be the object that handles login state.
Changed LabelAndTextFieldFactory to PanelFactory, which will be getting
more factory methods added to it because the forms will be giant bears