Big update
Cats can now be saved to the fake database.  There are a number of bugs
that need to be resolved. Namely, the image isn't saving to the cat
object, although that may need to be fixed on the database end. There is
also a couple bugs with the Login Dialog that are of high priority.