Started MedicalHistory in AnimalInfoView
Started work on adding medical history view to the AnimalInfoView. What
remains is to add an option for adding lines to the view, creating
labels for each panel of data since the view is getting sufficiently
large, and to save the data out to the database.  Also would be nice to
add a text area for comments.